+++ apps/next/components/PrintProviderData.tsx
@@ -22,12 +22,14 @@ export function PrintProviderData() {
         Project URI:
         <RawDisplayer data={metadataSettings?.projectURI} />
       </div>
-      <div>
-        Name: <RawDisplayer data={tokenSettings?.[0]} />
-        Symbol: <RawDisplayer data={tokenSettings?.[1]} />
-        Total Supply:
-        <RawDisplayer data={tokenSettings?.[2].toString()} />
-      </div>
+      {tokenSettings ? (
+        <div>
+          Name: <RawDisplayer data={tokenSettings?.[0]} />
+          Symbol: <RawDisplayer data={tokenSettings?.[1]} />
+          Total Supply:
+          {/* <RawDisplayer data={tokenSettings?.[2].toString()} /> */}
+        </div>
+      ) : null}
     </>
   )
 }